+++ Cargo.toml
@@ -7,5 +7,8 @@ members = [
 [profile.dev.package.backtrace]
 opt-level = 3
 
+[profile.dev.package.num-bigint-dig]
+opt-level = 3
+
 [profile.release]
 debug = true
