+++ crates/github_scbot_crypto/Cargo.toml
@@ -7,5 +7,6 @@ edition = "2018"
 [dependencies]
 jsonwebtoken = "7"
 openssl = "0.10"
+rsa = "0.5.0"
 serde = { version = "1", features = ["derive"] }
 thiserror = "1.0"
